pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "trackinsight-data-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "polars",
    "python-dotenv",
    "python-dateutil",
    "requests",
]

[project.optional-dependencies]
test = ["pytest>=8.0"]

[project.scripts]
trackinsight-data-python = "trackinsight_data_python:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
